backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
SQLAlchemy
geoalchemy2
pandas
scikit-learn
joblib
numpy
python-dotenv
geopy
openmeteo-requests
requests-cache
retry-requests